Update deploy.ps1
Added new check to see if the user which is connected to Azure is the same as the service account. If the accounts are different assign both the service account and the account logged on to Azure permissions. If the accounts are the same only assign the permissions to the service account